requirements.txt
claude-code-sdk>=0.0.25,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0.0

# E2B sandbox execution
e2b>=2.2.0,<3.0.0
e2b-desktop>=2.2.0,<3.0.0

# MCP protocol (if not already included by claude-code-sdk)
mcp>=0.1.0,<2.0.0

# Concurrency safety
filelock>=3.0.0,<4.0.0

# TUI interface
rich>=13.0.0,<15.0.0

# Vector embeddings for semantic similarity (optional but recommended)
sentence-transformers>=2.2.0,<6.0.0
numpy>=1.24.0,<3.0.0
